refactor: extract the four view panels out of App.tsx (move-only)
App.tsx was a 1542-line god-file. Moved the four self-contained,
props-only view components into src/components/, with zero behaviour
change:

- src/components/HomePanel.tsx   (+ HOME_CONTENT_STATS moved alongside)
- src/components/LearningPanel.tsx
- src/components/RulesPanel.tsx
- src/components/MockExamPanel.tsx

App.tsx 1542 -> ~990 lines. These four take everything via props and
don't touch App's session state, so the extraction is a pure move:
App now imports + renders them exactly as before. Dropped the imports
that only they used (CONJUGATION_TABLES, getMockExamBlueprint,
isLearningBlockComplete, getIncompletePrereqs, learningBlocks,
ClipboardList, BooksSpot, BrushSpot, type Copy).

Deferred to follow-up PRs (they share the Feedback type + session
state, so they need a shared types module / hook first):
- leaf components ExamPrompt / FeedbackPanel / SpeakButton
- the App core / ChallengePanel + usePracticeSession / useTheme hooks

Behaviour-identical: the App.test integration tests render the full app
and click through all four panels (nav tabs, chapter buttons, mock
section picker) -- all 118 pass. tsc + build + check:exam (303/0) green.
