+++ src/App.tsx
@@ -4,7 +4,6 @@ import {
   ArrowRight,
   BookOpen,
   CheckCircle2,
-  ClipboardList,
   Eye,
   GraduationCap,
   Moon,
@@ -17,14 +16,8 @@ import { ADJECTIVE_FORMS, VERB_FORMS } from "./domain/conjugation";
 import { buildClozeQuestionPool } from "./domain/cloze";
 import { clozeSentences } from "./domain/cloze-data";
 import { buildExamQuestionPool } from "./domain/examBlocks";
-import { getMockExamBlueprint, type MockExamLevel } from "./domain/mockExam";
-import { CONJUGATION_TABLES } from "./domain/conjugationTables";
-import {
-  getIncompletePrereqs,
-  isLearningBlockComplete,
-  learningBlocks,
-  type LearningBlockDrillPreset
-} from "./domain/learningBlocks";
+import type { MockExamLevel } from "./domain/mockExam";
+import type { LearningBlockDrillPreset } from "./domain/learningBlocks";
 import { buildSentencePatternPool, type SentencePatternId } from "./domain/sentencePatterns";
 import {
   buildChoiceOptions,
@@ -41,8 +34,12 @@ import { readStored, writeStored } from "./domain/safeStorage";
 import type { Attempt, PartOfSpeech, PracticeQuestion, TargetForm, VerbGroup } from "./domain/types";
 import { vocabulary } from "./domain/vocabulary";
 import { jlptVocabulary } from "./domain/vocabulary-jlpt";
-import { copy, type Copy, type Language } from "./i18n";
-import { BooksSpot, BrushSpot, DarumaSpot, PaperNoteSpot, TeaCupSpot } from "./illustrations";
+import { copy, type Language } from "./i18n";
+import { DarumaSpot, PaperNoteSpot, TeaCupSpot } from "./illustrations";
+import { HomePanel } from "./components/HomePanel";
+import { LearningPanel } from "./components/LearningPanel";
+import { RulesPanel } from "./components/RulesPanel";
+import { MockExamPanel } from "./components/MockExamPanel";
 import "./styles.css";
 
 type Feedback =
@@ -830,401 +827,6 @@ export default function App() {
   );
 }
 
-function LearningPanel({
-  language,
-  progressAttempts,
-  reviewCount,
-  onStartChallenge,
-  onStartReview,
-  onStartDrill,
-  onStartPatternDrill
-}: {
-  language: Language;
-  progressAttempts: Attempt[];
-  reviewCount: number;
-  onStartChallenge: () => void;
-  onStartReview: () => void;
-  onStartDrill: (preset: DrillPreset) => void;
-  onStartPatternDrill: (patternIds: SentencePatternId[]) => void;
-}) {
-  const t = copy[language];
-  // PR A: surface only the "basic" learning blocks. PR C will introduce
-  // exam-prep blocks alongside these.
-  const blockCards = learningBlocks
-    .filter((block) => block.group === "basic")
-    .map((block) => ({
-      block,
-      complete: isLearningBlockComplete(progressAttempts, block),
-      incompletePrereqs: getIncompletePrereqs(progressAttempts, block)
-    }));
-
-  // Default to the first incomplete chapter; fall back to the first
-  // chapter if everything is complete (e.g. revisiting after finishing).
-  const recommended = blockCards.find((card) => !card.complete) ?? blockCards[0];
-
-  const [selectedBlockId, setSelectedBlockId] = useState<string | null>(null);
-  const activeCard = blockCards.find((card) => card.block.id === selectedBlockId) ?? recommended;
-  const active = activeCard.block;
-
-  // Resolve the drill button label from the i18n copy table. The schema
-  // stores a plain string key so new drill labels don't require schema
-  // updates.
-  const drillButtonLabel = (drill: { labelKey: string }): string => {
-    const labels = t as unknown as Record<string, string>;
-    return labels[drill.labelKey] ?? drill.labelKey;
-  };
-
-  const blockTitleById = (id: string): string => {
-    const found = learningBlocks.find((b) => b.id === id);
-    return found ? found.title : id;
-  };
-
-  // Lightweight dashboard stats: derived from the same attemptStore the
-  // chapter index uses, so the count stays in sync with whatever the
-  // learner just did. accuracyPercent is computed across ALL recorded
-  // attempts (not just this session) -- the dashboard's purpose is to
-  // surface cross-session signal, not session-local feedback.
-  const dashboardTotalAttempts = progressAttempts.length;
-  const dashboardCorrectAttempts = progressAttempts.filter((a) => a.isCorrect).length;
-  const dashboardAccuracy =
-    dashboardTotalAttempts === 0
-      ? 0
-      : Math.round((dashboardCorrectAttempts / dashboardTotalAttempts) * 100);
-
-  return (
-    <section className="learning-panel" aria-label={t.learningRegion}>
-      <div className="chapter-shell">
-        <aside className="chapter-index" aria-label="學習章節">
-          <div className="dashboard-card" aria-label={t.dashboardEyebrow}>
-            <p className="eyebrow">{t.dashboardEyebrow}</p>
-            {reviewCount > 0 ? (
-              <button
-                type="button"
-                className="dashboard-review-cta"
-                onClick={onStartReview}
-              >
-                <AlertTriangle aria-hidden="true" />
-                <span className="dashboard-review-text">
-                  {t.dashboardReviewPending(reviewCount)}
-                </span>
-                <span className="dashboard-review-action">{t.dashboardReviewCta}</span>
-              </button>
-            ) : (
-              <p className="dashboard-review-empty">{t.dashboardReviewEmpty}</p>
-            )}
-            {dashboardTotalAttempts > 0 ? (
-              <div className="dashboard-stats">
-                <span>{t.dashboardStatsAttempts(dashboardTotalAttempts)}</span>
-                <span>·</span>
-                <span>{t.dashboardStatsAccuracy(dashboardAccuracy)}</span>
-              </div>
-            ) : null}
-          </div>
-          <div className="chapter-index-copy">
-            <p className="eyebrow">課程章節</p>
-            <h2>一章一章解鎖</h2>
-            <p>選一章看規則、例子與常見陷阱，再到挑戰頁練。</p>
-          </div>
-
-          <div className="chapter-list">
-            {blockCards.map(({ block, complete, incompletePrereqs }) => (
-              <button
-                key={block.id}
-                type="button"
-                className={`chapter-list-button${block.id === active.id ? " selected" : ""}${complete ? " complete" : ""}`}
-                aria-label={`查看：${block.title}`}
-                aria-pressed={block.id === active.id}
-                onClick={() => setSelectedBlockId(block.id)}
-              >
-                <span>
-                  {block.completionMode === "reference"
-                    ? "參考"
-                    : complete
-                    ? "完成"
-                    : block.kicker ?? block.category}
-                </span>
-                <strong>{block.title}</strong>
-                <small>
-                  {incompletePrereqs.length > 0
-                    ? `建議先看：${incompletePrereqs.map(blockTitleById).join("、")}`
-                    : block.subtitle}
-                </small>
-              </button>
-            ))}
-          </div>
-
-          <button className="secondary-challenge-button" type="button" onClick={onStartChallenge}>
-            <ArrowRight aria-hidden="true" />
-            {t.startChallenge}
-          </button>
-        </aside>
-
-        <section className="chapter-content" aria-labelledby="active-chapter-title">
-          <div className="chapter-content-head">
-            <p className="eyebrow">{active.kicker ?? active.category}</p>
-            <h3 id="active-chapter-title">{active.title}</h3>
-            <p>{active.explanation}</p>
-            {active.subtitle ? (
-              <div className="focus-formula" aria-label={`${active.title}例子`}>
-                <span>{active.subtitle}</span>
-              </div>
-            ) : null}
-          </div>
-
-          <div className="chapter-lesson">
-            <div className="pipeline-grid">
-              {active.examples.map((example) => (
-                <article className="pipeline-card" key={example.formula}>
-                  <code>{example.formula}</code>
-                  {example.note ? <p>{example.note}</p> : null}
-                </article>
-              ))}
-            </div>
-
-            {active.pitfalls && active.pitfalls.length > 0 ? (
-              <div className="block-pitfalls">
-                <p className="block-pitfalls-title">常見陷阱</p>
-                <ul>
-                  {active.pitfalls.map((pitfall) => (
-                    <li key={pitfall}>{pitfall}</li>
-                  ))}
-                </ul>
-              </div>
-            ) : null}
-
-            {active.patternDrills && active.patternDrills.length > 0 ? (
-              <div className="inline-action-row">
-                {active.patternDrills.map((drill) => (
-                  <button
-                    key={drill.labelKey}
-                    className="inline-drill-button"
-                    type="button"
-                    onClick={() => onStartPatternDrill(drill.patternIds)}
-                  >
-                    <ArrowRight aria-hidden="true" />
-                    {drillButtonLabel(drill)}
-                  </button>
-                ))}
-              </div>
-            ) : null}
-
-            {active.drillNote ? (
-              <p className="block-drill-note">{active.drillNote}</p>
-            ) : null}
-
-            {active.drills && active.drills.length > 0 ? (
-              <div className="inline-action-row">
-                {active.drills.map((drill) => (
-                  <button
-                    key={drill.labelKey}
-                    className="inline-drill-button"
-                    type="button"
-                    onClick={() => onStartDrill(drill.preset)}
-                  >
-                    <ArrowRight aria-hidden="true" />
-                    {drillButtonLabel(drill)}
-                  </button>
-                ))}
-              </div>
-            ) : null}
-          </div>
-        </section>
-      </div>
-    </section>
-  );
-}
-
-// ---- HomePanel -------------------------------------------------------
-// First view the learner lands on. Three layers:
-//   1. Context-aware banner ("review N items" if any, else "continue
-//      chapter XX" if there's an incomplete one). Suppressed entirely
-//      when nothing meaningful to surface, to avoid noise.
-//   2. Lifetime stats strip (only shown after the first attempt).
-//   3. Four entry cards. The Review card disables itself when there's
-//      nothing to review, so the affordance is still discoverable but
-//      doesn't lead to an empty quiz.
-//
-// HomePanel is intentionally read-only of the learner state -- mutating
-// callbacks (onNavigate, onStartReview) live on the parent so the panel
-// stays a presentational component.
-
-// Content-volume snapshot rendered above the entry cards. Computed
-// once at module load -- the underlying data (learningBlocks, exam
-// pool, sentence patterns, jlptVocabulary) is static at runtime and
-// only changes when a content batch ships, which rebuilds the bundle
-// anyway. Pre-computing avoids re-running buildExamQuestionPool on
-// every HomePanel render.
-const HOME_CONTENT_STATS = {
-  chapters: learningBlocks.filter((block) => block.group === "basic").length,
-  examItems: buildExamQuestionPool("all").length,
-  n1Grammar: buildExamQuestionPool("N1").filter((q) => q.promptLabel === "文法形式選擇").length,
-  patternChecks: buildSentencePatternPool().length,
-  vocab: jlptVocabulary.length
-};
-
-function HomePanel({
-  language,
-  progressAttempts,
-  reviewCount,
-  onNavigate,
-  onStartReview,
-  onStartVocab
-}: {
-  language: Language;
-  progressAttempts: Attempt[];
-  reviewCount: number;
-  onNavigate: (target: "learn" | "challenge" | "mock") => void;
-  onStartReview: () => void;
-  onStartVocab: () => void;
-}) {
-  const t = copy[language];
-
-  const totalAttempts = progressAttempts.length;
-  const correctAttempts = progressAttempts.filter((attempt) => attempt.isCorrect).length;
-  const accuracyPercent =
-    totalAttempts === 0 ? 0 : Math.round((correctAttempts / totalAttempts) * 100);
-
-  // Only count "trackable" basic chapters towards the X / Y badge --
-  // reference chapters (verb-types + the four sentence-pattern reading
-  // chapters) are reading material, not drillable units.
-  const trackableChapters = learningBlocks.filter(
-    (block) => block.group === "basic" && block.completionMode !== "reference"
-  );
-  const completedChapters = trackableChapters.filter((block) =>
-    isLearningBlockComplete(progressAttempts, block)
-  ).length;
-
-  const nextIncompleteChapter = trackableChapters.find(
-    (block) => !isLearningBlockComplete(progressAttempts, block)
-  );
-
-  return (
-    <section className="home-panel" aria-label={t.home}>
-      <header className="home-hero">
-        {/* Decorative hero -- the heading below carries the actual
-            message, so alt is intentionally empty (avoids the screen
-            reader saying "image, hero" which adds nothing). */}
-        <img
-          className="home-hero-image"
-          src="/hero.webp"
-          alt=""
-          width={1600}
-          height={900}
-        />
-        <div className="home-hero-text">
-          <h1>{t.homeHeroTitle}</h1>
-          <p>{t.homeHeroIntro}</p>
-        </div>
-      </header>
-
-      {/* Content-volume strip: tells first-time visitors what they're
-          walking into without resorting to SaaS-style metric tiles.
-          Counts are derived live from the data modules so this stays
-          honest whenever a content batch lands. */}
-      <p className="home-content-stats">
-        {t.homeContentStats(
-          HOME_CONTENT_STATS.chapters,
-          HOME_CONTENT_STATS.examItems,
-          HOME_CONTENT_STATS.n1Grammar,
-          HOME_CONTENT_STATS.patternChecks,
-          HOME_CONTENT_STATS.vocab
-        )}
-      </p>
-
-      {reviewCount > 0 ? (
-        <button type="button" className="home-banner home-banner-review" onClick={onStartReview}>
-          <AlertTriangle aria-hidden="true" />
-          <span className="home-banner-text">
-            <strong>{t.homeBannerReviewMain(reviewCount)}</strong>
-            <small>{t.homeBannerReviewSub}</small>
-          </span>
-          <ArrowRight aria-hidden="true" />
-        </button>
-      ) : nextIncompleteChapter ? (
-        <button
-          type="button"
-          className="home-banner home-banner-continue"
-          onClick={() => onNavigate("learn")}
-        >
-          <BookOpen aria-hidden="true" />
-          <span className="home-banner-text">
-            <strong>{t.homeBannerContinueMain(nextIncompleteChapter.title)}</strong>
-            <small>{t.homeBannerContinueSub}</small>
-          </span>
-          <ArrowRight aria-hidden="true" />
-        </button>
-      ) : null}
-
-      {totalAttempts > 0 ? (
-        <div className="home-stats-strip" aria-label={t.homeStatsLabel}>
-          <div className="home-stats-cell">
-            <strong>{totalAttempts}</strong>
-            <small>{t.homeStatsAttempts}</small>
-          </div>
-          <div className="home-stats-cell">
-            <strong>{accuracyPercent}%</strong>
-            <small>{t.homeStatsAccuracy}</small>
-          </div>
-          <div className="home-stats-cell">
-            <strong>
-              {completedChapters} / {trackableChapters.length}
-            </strong>
-            <small>{t.homeStatsChapters}</small>
-          </div>
-        </div>
-      ) : null}
-
-      <div className="home-grid">
-        {/* Each card carries a single-CJK "stage badge" (學 / 練 / 背 /
-            考 / 補) instead of a lucide-react icon. The icons read as
-            tech-product chrome; the kanji read as editorial. Stages
-            suggest a natural progression but the cards stay
-            independently entry-able -- a returning learner who only
-            wants today's mock exam can still jump straight to 考. */}
-        <button type="button" className="home-card" onClick={() => onNavigate("learn")}>
-          <span className="home-card-stage" aria-hidden="true">{t.homeCardStageLearn}</span>
-          <h2>{t.homeCardLearnTitle}</h2>
-          <p>{t.homeCardLearnSub}</p>
-          <span className="home-card-meta">
-            {t.homeCardLearnMeta(completedChapters, trackableChapters.length)}
-          </span>
-          <ArrowRight className="home-card-arrow" aria-hidden="true" />
-        </button>
-        <button type="button" className="home-card" onClick={() => onNavigate("challenge")}>
-          <span className="home-card-stage" aria-hidden="true">{t.homeCardStageChallenge}</span>
-          <h2>{t.homeCardChallengeTitle}</h2>
-          <p>{t.homeCardChallengeSub}</p>
-          <span className="home-card-meta">{t.homeCardChallengeMeta}</span>
-          <ArrowRight className="home-card-arrow" aria-hidden="true" />
-        </button>
-        <button type="button" className="home-card" onClick={onStartVocab}>
-          <span className="home-card-stage" aria-hidden="true">{t.homeCardStageVocab}</span>
-          <h2>{t.homeCardVocabTitle}</h2>
-          <p>{t.homeCardVocabSub}</p>
-          <span className="home-card-meta">{t.homeCardVocabMeta}</span>
-          <ArrowRight className="home-card-arrow" aria-hidden="true" />
-        </button>
-        <button type="button" className="home-card" onClick={() => onNavigate("mock")}>
-          <span className="home-card-stage" aria-hidden="true">{t.homeCardStageMock}</span>
-          <h2>{t.homeCardMockTitle}</h2>
-          <p>{t.homeCardMockSub}</p>
-          <span className="home-card-meta">{t.homeCardMockMeta}</span>
-          <ArrowRight className="home-card-arrow" aria-hidden="true" />
-        </button>
-        <button type="button" className="home-card" onClick={onStartReview}>
-          <span className="home-card-stage" aria-hidden="true">{t.homeCardStageReview}</span>
-          <h2>{t.homeCardReviewTitle}</h2>
-          <p>
-            {reviewCount > 0 ? t.homeCardReviewSubActive(reviewCount) : t.homeCardReviewSubEmpty}
-          </p>
-          <span className="home-card-meta">{t.homeCardReviewMeta}</span>
-          <ArrowRight className="home-card-arrow" aria-hidden="true" />
-        </button>
-      </div>
-    </section>
-  );
-}
-
 function getInitialTheme(): Theme {
   const storedTheme = readStored(THEME_STORAGE_KEY);
 
@@ -1243,57 +845,6 @@ function getInitialTheme(): Theme {
 // Pure presentational; all data lives in domain/conjugationTables.ts
 // so future tables (formal forms / adjective variations / sentence
 // patterns) just need a data-file edit and no component change.
-function RulesPanel({ language }: { language: Language }) {
-  const t = copy[language];
-  return (
-    <section className="rules-panel" aria-label={t.rules}>
-      <header className="rules-header">
-        <BrushSpot className="panel-header-spot" />
-        <p className="eyebrow">{t.rulesEyebrow}</p>
-        <h1>{t.rulesPanelTitle}</h1>
-        <p>{t.rulesPanelIntro}</p>
-      </header>
-      {CONJUGATION_TABLES.map((table) => (
-        <article key={table.id} className="rules-table-card">
-          <header className="rules-table-head">
-            <h2>{table.title}</h2>
-            <p>{table.caption}</p>
-          </header>
-          <div className="rules-table-wrap" role="region" aria-label={table.title}>
-            <table>
-              <thead>
-                <tr>
-                  {table.columns.map((col) => (
-                    <th key={col} scope="col">
-                      {col}
-                    </th>
-                  ))}
-                </tr>
-              </thead>
-              <tbody>
-                {table.rows.map((row, rowIndex) => (
-                  <tr key={rowIndex}>
-                    {row.map((cell, cellIndex) => (
-                      <td key={cellIndex}>{cell}</td>
-                    ))}
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-          {table.pitfalls && table.pitfalls.length > 0 ? (
-            <ul className="rules-pitfalls">
-              {table.pitfalls.map((pitfall) => (
-                <li key={pitfall}>{pitfall}</li>
-              ))}
-            </ul>
-          ) : null}
-        </article>
-      ))}
-    </section>
-  );
-}
-
 function storeTheme(theme: Theme) {
   writeStored(THEME_STORAGE_KEY, theme);
 }
@@ -1437,106 +988,6 @@ function FeedbackPanel({ feedback, language }: { feedback: NonNullable<Feedback>
           <span>{feedback.question.vocabulary.examples[0].meaningZh}</span>
         </p>
       ) : null}
-    </section>
-  );
-}
-
-// ---- Mock exam panel (section picker) ----------------------------------
-// Reworked from a strict timed full-paper exam into a section picker: the
-// learner taps a JLPT section and drills just that section in the normal
-// challenge view (with per-question feedback + SRS recording). Less
-// rigorous than a timed paper, and -- unlike the old flow -- it always
-// works because it just hands a section filter to the existing exam pool.
-// The blueprint (mockExam.ts) still supplies the canonical section labels
-// and order; sections with no authored items yet are shown disabled.
-
-function MockExamPanel({
-  language,
-  onStartSection
-}: {
-  language: Language;
-  onStartSection: (level: MockExamLevel, promptLabel: string) => void;
-}) {
-  const t = copy[language];
-  const [level, setLevel] = useState<MockExamLevel>("N2");
-
-  const blueprint = getMockExamBlueprint(level);
-  const pool = buildExamQuestionPool(level);
-  // Live count of available questions per section (keyed by promptLabel).
-  const counts = new Map<string, number>();
-  for (const question of pool) {
-    const key = question.promptLabel ?? "";
-    counts.set(key, (counts.get(key) ?? 0) + 1);
-  }
-
-  return (
-    <section className="mock-panel" aria-label={t.mockExam}>
-      <header className="mock-section-head">
-        <BooksSpot className="panel-header-spot" />
-        <p className="eyebrow">
-          <ClipboardList aria-hidden="true" />
-          {t.mockSectionTitle}
-        </p>
-        <p className="mock-section-intro">{t.mockSectionIntro}</p>
-      </header>
-
-      <fieldset className="mock-level-picker">
-        <legend>{t.mockExamLevelLabel}</legend>
-        <div className="segmented">
-          {(["N2", "N1"] as MockExamLevel[]).map((option) => (
-            <button
-              key={option}
-              type="button"
-              className={level === option ? "selected" : ""}
-              onClick={() => setLevel(option)}
-            >
-              {option}
-            </button>
-          ))}
-        </div>
-      </fieldset>
-
-      <ol className="mock-section-rows">
-        {blueprint.sections.map((section) => {
-          const count = counts.get(section.promptLabel) ?? 0;
-          // Sections with no authored questions yet render as a plain
-          // info row ("準備中"), NOT a disabled button -- you can't drill
-          // an empty pool, so it shouldn't look like a dead action.
-          // Sections with questions are clickable cards.
-          if (count === 0) {
-            return (
-              <li key={section.id}>
-                <div className="mock-section-card empty">
-                  <div className="mock-section-meta">
-                    <strong>{section.labelJa}</strong>
-                    <small>{section.labelZh}</small>
-                  </div>
-                  <span className="mock-section-warn">
-                    <AlertTriangle aria-hidden="true" />
-                    {t.mockSectionEmpty}
-                  </span>
-                </div>
-              </li>
-            );
-          }
-          return (
-            <li key={section.id}>
-              <button
-                type="button"
-                className="mock-section-card"
-                onClick={() => onStartSection(level, section.promptLabel)}
-              >
-                <div className="mock-section-meta">
-                  <strong>{section.labelJa}</strong>
-                  <small>{section.labelZh}</small>
-                </div>
-                <span className="mock-section-count">{t.mockSectionCount(count)}</span>
-                <ArrowRight className="mock-section-arrow" aria-hidden="true" />
-              </button>
-            </li>
-          );
-        })}
-      </ol>
     </section>
   );
 }
